# docs/swagger.yaml
basePath: /v1
definitions:
  domainpayment.PaymentType:
    enum:
    - CASH
    - E-WALLET
    - EDC
    type: string
    x-enum-varnames:
    - Cash
    - EWallet
    - EDC
...
  modelv1.AuthResponse:
    properties:
      token:
        example: v2.local.Gdh5kiOTyyaQ3_bNykYDeYHO21Jg2...
        type: string
    type: object
  modelv1.CategoryResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Foods
        type: string
    type: object
  modelv1.CreateCategoryRequest:
    properties:
      name:
        example: Foods
        type: string
    required:
    - name
    type: object
  modelv1.CreateOrderRequest:
    properties:
      customer_name:
        example: John Doe
        type: string
      payment_id:
        example: 1
        type: integer
      products:
        items:
          $ref: '#/definitions/modelv1.OrderProductRequest'
        type: array
      total_paid:
        example: 100000
        type: integer
    required:
    - customer_name
    - payment_id
    - products
    - total_paid
    type: object
...
  modelv1.ErrorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
  modelv1.LoginRequest:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: "12345678"
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  modelv1.Meta:
    properties:
      limit:
        example: 10
        type: integer
      skip:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
  modelv1.OrderProductRequest:
    properties:
      product_id:
        example: 1
        minimum: 1
        type: integer
      qty:
        example: 1
        type: integer
    required:
    - product_id
    - qty
    type: object
  modelv1.OrderProductResponse:
    properties:
      created_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      order_id:
        example: 1
        type: integer
      price:
        example: 100000
        type: number
      product:
        $ref: '#/definitions/modelv1.ProductResponse'
      product_id:
        example: 1
        type: integer
      qty:
        example: 1
        type: integer
      total_final_price:
        example: 100000
        type: number
      total_normal_price:
        example: 100000
        type: number
      updated_at:
        example: "1970-01-01T00:00:00Z"
        type: string
    type: object
  modelv1.OrderResponse:
    properties:
      created_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      customer_name:
        example: John Doe
        type: string
      id:
        example: 1
        type: integer
      payment_type:
        $ref: '#/definitions/modelv1.PaymentResponse'
      payment_type_id:
        example: 1
        type: integer
      products:
        items:
          $ref: '#/definitions/modelv1.OrderProductResponse'
        type: array
      receipt_id:
        example: 4979cf6e-d215-4ff8-9d0d-b3e99bcc7750
        type: string
      total_paid:
        example: 100000
        type: number
      total_price:
        example: 100000
        type: number
      total_return:
        example: 0
        type: number
      updated_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  modelv1.PaymentResponse:
    properties:
      id:
        example: 1
        type: integer
      logo:
        example: https://example.com/cash.png
        type: string
      name:
        example: Tunai
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domainpayment.PaymentType'
        example: CASH
    type: object
  modelv1.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/modelv1.CategoryResponse'
      created_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/chiki-ball.png
        type: string
      name:
        example: Chiki Ball
        type: string
      price:
        example: 5000
        type: number
      sku:
        example: 9a4c25d3-9786-492c-b084-85cb75c1ee3e
        type: string
      stock:
        example: 100
        type: integer
      updated_at:
        example: "1970-01-01T00:00:00Z"
        type: string
    type: object
...
  modelv1.Response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
  modelv1.UpdateCategoryRequest:
    properties:
      name:
        example: Beverages
        type: string
    required:
    - name
    type: object
...
host: localhost
info:
  contact:
    email: letienminh2512@gmail.com
    name: Tien Minh
    url: https://github.com/TienMinh25/go-hexagonal-architecture
  description: This is a simple RESTful Point of Sale (POS) Service API written in
    Go using Gin web framework, PostgreSQL database, and Redis cache.
  license:
    name: APACHE
    url: https://github.com/TienMinh25/go-hexagonal-architecture/blob/main/LICENSE
  title: Go Sale API demo for hexagonal architecture
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: List categories with pagination
      parameters:
      - description: Skip
        in: query
        name: skip
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Categories displayed
          schema:
            $ref: '#/definitions/modelv1.Meta'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: create a new category with name
      parameters:
      - description: Create category request
        in: body
        name: createCategoryRequest
        required: true
        schema:
          $ref: '#/definitions/modelv1.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category created
          schema:
            $ref: '#/definitions/modelv1.CategoryResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted
          schema:
            $ref: '#/definitions/modelv1.Response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: get a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category retrieved
          schema:
            $ref: '#/definitions/modelv1.CategoryResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: update a category's name by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update category request
        in: body
        name: updateCategoryRequest
        required: true
        schema:
          $ref: '#/definitions/modelv1.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated
          schema:
            $ref: '#/definitions/modelv1.CategoryResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
  /orders:
    get:
      consumes:
      - application/json
      description: List orders and return an array of order data with purchase details
      parameters:
      - description: Skip records
        in: query
        name: skip
        required: true
        type: integer
      - description: Limit records
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders displayed
          schema:
            $ref: '#/definitions/modelv1.Meta'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order and return the order data with purchase details
      parameters:
      - description: Create order request
        in: body
        name: createOrderRequest
        required: true
        schema:
          $ref: '#/definitions/modelv1.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order created
          schema:
            $ref: '#/definitions/modelv1.OrderResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get an order by id and return the order data with purchase details
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order displayed
          schema:
            $ref: '#/definitions/modelv1.OrderResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an order
      tags:
      - Orders
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Logs in a registered user and returns an access token if the credentials
        are valid.
      parameters:
      - description: Login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modelv1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Succesfully logged in
          schema:
            $ref: '#/definitions/modelv1.AuthResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      summary: Login and get an access token
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"